unify all ivy"x::y::z" dependencies JVM/JS/Native
There is still some confusion about scalajs dependencies (mill have different notation than sbt in this place). It's not well documented what needs to be changed to use scalajs dependencies on ScalaJSModule. 

On the other hand it is safe to say "always use smart ivy dependency" which'll always chose proper one. Doc can have examples with this assumption in it.