Add support for Dotty projects (#397)
* Abstract over the scala compiler organization

* Support using a locally published compiler

Publishing locally with sbt means publishing ivy-style, which uses
a different naming convention than maven, we now handle both cases.

* Add minimal support for Dotty projects

* Rewrite scalalib.Dep, introduce scalalib.CrossVersion

Instead of Dep being a trait with three cases (Java/Scala/Point), it is
now a case class where the cross field is an instance of the
CrossVersion trait which has three cases (Constant/Binary/Full). This is
more versatile since it allows for non-empty constant suffixes which
will be used to implement withDottyCompat in the next commit. It's
also a cleaner separation of concerns. We also deduplicate various
pieces of codes that computed the artifact name: this is now always handled in
Dep and CrossVersion.

* Add simple way to use Scala 2 deps in a Dotty project

This is similar to the withDottyCompat method in the sbt-dotty plugin.

* Turn off the Dotty test on Java >= 9