Add task to only check formatting of Scala files (#733)
Adds a new task 'checkFormat' to the scalafmt module that only checks if all files are formatted. It is basically equivalent to the --test flag that can be passed to scalafmt on the command line, and can come in very handy in CI scripts.

Caching is preserved wrt to the 'reformat' action, so that only unformatted or changed files are checked.