Add support for Scala.js 0.6.29+ and 1.0.0-RC1.
Drop support for the obsolete Scala.js 1.0.0-M2.

Starting from Scala.js 0.6.29, test artifacts must depend on
`scalajs-test-bridge` instead of `scalajs-test-interface`, as the
JS bridge has been moved to that artifact. This is the only change
required for 0.6.29+.

In Scala.js 1.x, in addition to depending on
`scalajs-test-bridge`, test artifacts must include a module
initializer that initializes the bridge. This required the
addition of a `testBridgeInit: Boolean` parameter to the `link`
methods scattered here and there. The parameter is ignored in 0.6.

The support for 1.0.0-RC1 is suboptimal: we need to force emitting
ECMAScript 5.1 code instead of the new default ES 2015. The only
reason is that the tests of Mill itself rely on Nashorn to execute
some Scala.js outputs, and Nashorn does not support ES 2015. Since
there was no configuration mechanism for the ES mode, I decided to
always force ES 5.1. It would be better to introduce a
configuration option for that in the future.