Replaced `git clean` with `git stash` in ci script
The effect is the same (all untrack files will be deleted)
But in contrast to `git clean` were all files will be deleted
in `git stash` all untrack files will be stashed and can be
restored, in case of an accidental execution of a ci script in
and developer repo.

We call two `git stash` commands:
1. `git stash -u` to stash all untrack files, but not those
   which are also ignored
2. `git stash -a` to stash all other files, even those which
   are ignored by git to make a really clean work dir.

The purpose of these 2 stash entries is, that it will be
easier to examine and restore potential important files
(most probably in the first stash) from generated and ignored
files (probably in the second stash).

# Conflicts:
#	ci/test-mill-0.sh