Relative library paths in GenIdea, and a relative COURSIER_CACHE on WSL2 (#887)
This should allow the GenIdea output XML files to be usable cross-platform, e.g. we could resolve them on a WSL-Ubuntu environment and load them into Intellij on the host Windows machine

Tested locally on OS-X, seems to work, and tested on WSL2 as well. IntelliJ seems to be able to load the config successfully, even across the Ubuntu/Windows boundary