Replace scala-library artifact with dotty-library (#915)
This change is a "symptomatic treatment" to fix bloop config
generation, which is currently broken for dotty projects.

In particular, [this
issue](https://github.com/scalameta/metals/issues/1768)
describes the error and also hints to the fix.

The gist of it seems to be that Dotty publishes an empty jar named
"scala-library" (not sure why?), but the actual standard library is
published in the artifact "dotty-library". While the former is empty
and isn't strictly a problem when compiling with dotty, it does
cause bloop to crash if included in the classpath.

Note that I am not exactly sure what the root cause is, and I'm
also not sure what Dotty's future plans are regarding the naming of
artifacts, so it would be great to get feedback from someone more
knowledgeable in this area before accepting these changes.