Exclude trees from ammonite dep.
The reason for doing this is to ensure that the version of
trees that mill wants is indeed used instead of the newer
version that is being pulled in by ammonite-interp. Before
this was using `forceVersion()`, but by doing this the forced
version doesn't end up in the POM meaning that when someone
bootstraps/uses mill via coursier, they are getting the newer
trees, which is causing issues with how the main class is detected.

You can see more details on this https://github.com/coursier/apps/issues/56
and a huge props to @alexarchambault for finding this.