Prefix the ticker of evaluator targets with the caller target
This avoids the confusion regarding the jumping overall target count.
Now, we can see clearly that the jumping overall target count is a sub-count.

This result in an output like this:

```
$ mill mill.scalalib.GenIdea/idea
[1/1] idea
Analyzing modules ...
[1/1] idea > [1270/1271] app.headless.test.ideaCompileOutput
```

I also fixed the ticker progress for parallel builds, which was 0-based,
resulting is some (incorrect) output like this:

```
$ dev-mill -j 0 idea
Using experimental parallel evaluator with 4 threads
[#0] [0/1] idea
[#0] Analyzing modules ...
[#3] [0/1] idea > [1188/1271] module.contract.test.ideaConfigFiles
```