Clean up BSP logging output.
Previously, the MillBspLogger proxied the evaluator logger, which meant all
output went twice to stdout, once through the proxy and once through the BSP client
messages.

MillBspLogger now does not proxy a logger, instead it only writes to the BSP
client. It will write whatever is emitted to its streams, or written
via the Logger methods on it.

The logger no longer uses the ticker formatting or colors offered by PrintLogger.
This decoration is only suitable for the terminal, since e.g. color codes and navigation codes
are unlikely to be supported in an editor's log output.

To debug this, I've added a log file that logs the raw json output from
BSP/start. I've also moved the BSP client logging to the BSP/start dest
directory, so it no longer goes in .bsp/mill.log.

ProxyLogger is no longer a case class. It is intended for subclassing,
so it not suitable as a case class.