Add version range sources to CrossScalaModule
With this feature new sources directories will be generated
based on the set of values in `Cross[T]` to allow easier management
of projects with multiple Scala versions.
In particular there will be a `segment+` and `segment-` for every
version that correctly matches the version `<=` and `>=` based on
the numbers making the semantic version scheme.