play-contrib: Fixed classloading of worker module and cleanups
Fixed dependency resolution of worker classpath.
The collected classpath of the worker module was limited to
only use the worker module.
But this misses all transitive dependencies of the worker.
This is probably because of older limitiations in mill builds,
which also included the mill API itself.
Now, mill uses `compileModuleDeps` for the mill API and we have
clean and usable transitive module and worker classpaths.

Fixed use of mills utility classloader by settings the correct parent classloader.
Using the classloader of the module as parent makes the shared API modules
(which need to be already loaded by the mill build) visible.
The perviously used default was using the classloader which loaded the
mill build (and the ammonite script), which misses the play API module
(loaded via `$ivy` import).