Declutter out/ folder by flattening out/foo/ foo folders into their constituent files (#1588)
This PR modifies the Mill evaluator output format, so that the `foo/meta.json` file is now top level as `foo.json`, as are the `foo.dest/` folder and `foo.log` file, although the latter two are not created unless necessary. This helps makes it easier to browse the output folder and find what you want: most "boring" tasks which do a trivial computation no longer generate a `foo/` folder, making the "interesting" tasks or the sub-modules stand out.

# Why

For example. Given the simple `build.sc` below:

```scala
import mill._
import mill.scalalib._

object foo extends ScalaModule{
  def scalaVersion = "2.13.2"
  object bar extends ScalaModule{
    def scalaVersion = "2.13.2"
  }

  object qux extends ScalaModule{
    def scalaVersion = "2.13.2"
  }
}
```

With the status quo, it is quite hard to look at the out folder and distinguish important tasks like "compile" or submodules like "bar" and "qux" in a sea of folders holding trivial task output. This holds true both in the terminal or in a graphical browser like an IDE:

<img width="706" alt="Screenshot 2021-11-25 at 2 22 26 PM" src="https://user-images.githubusercontent.com/934140/143390500-88b9179b-0e33-4f5e-bbb5-a2f2623f3139.png">

<img width="262" alt="Screenshot 2021-11-25 at 2 22 50 PM" src="https://user-images.githubusercontent.com/934140/143390511-e183e460-e7d6-4bce-a217-c953cb3ba857.png">

With this PR, the "trivial" tasks and "important" tasks or submodules are clearly visible at a glance, making it much easier to navigate the output folder:




# Changes

- We already lazily create the `foo/dest/` folder, so the only change here is to lazily create the `foo/log` file when someone writes to the output logger.

- I modified `./mill clean` to take delete both `foo/` and `foo.json` paths

- Updated the docs to refer to the new output layout

<img width="682" alt="Screenshot 2021-11-26 at 9 44 28 AM" src="https://user-images.githubusercontent.com/934140/143514513-c1fb8e0a-4580-48a7-a882-bd62a6a720d1.png">


<img width="286" alt="Screenshot 2021-11-26 at 9 44 11 AM" src="https://user-images.githubusercontent.com/934140/143514505-ebdb655a-5ee7-4871-8771-5681898ea5f4.png">


# Misc

Currently, we end up with *both* a `foo.json` file and a `foo.dest/` folder or `foo.log` file in the cases where they are necessary. I had considered automatically swapping between the `foo.json` file and the old `foo/meta.json` output style, but I figured that would make it harder to predict where the `.json` output file would end up. Decided to go with the proposed arrangement of `foo.json` and `foo.dest`/`foo.log` side by side for simplicity and predictability, even if it does make the output folder a bit more cluttered.

`rm -rf out/path/to/task/` no longer works for cleaning it; you'll have to remove the `out/path/to/task.*` now, or use `./mill clean`

Note that this has a chance of making the out folder even *more* cluttered, if most tasks have `.dest` and `.log` files defined. However, empirically it seems those tasks are very much the minority, and the vast majority of tasks are "trivial" tasks that only generate the `.json` file and nothing else


This is a big-ish change, so shouldn't be in a point version bump. Maybe it can go into 0.10.x?