Changed out path for overridden targets for consistency
To be more consistent with the changes from PR
https://github.com/com-lihaoyi/mill/pull/1588
which re-layouted the log, meta, and dest path for targets.
This change touches the path of overriden target results.
It replaces the `../target/overriden/..` part with
`../target.overriden/..`.

There is a catch though, to make this work I had to remove one
`require` assertion, which guarded the presence of dots (`.`) in
labels. TBH, I previously stubled upon this require
and also added the FIXME comment to it, as it is unclear why it
even exists. Probably, because labels with dots can't be properly
parsed when given on cli. More details and hisory can be found in
issue https://github.com/com-lihaoyi/mill/issues/820.

Fix https://github.com/com-lihaoyi/mill/issues/820.

Review by @lihaoyi