Explicit handling of scalac plugins and related enablement options (#1705)
We no longer generate an enablement option for each transitively resolved plugin jar, but instead, only generated one `-Xplugin:` option for each explicitly given plugin. This will fix previously seen warnings for the scalac compiler.

I took the opportunity to clean things up on the go.
I removed the handling of `-Xplugin:` from the ZincWorker into the ScalaModule.
This fixes the issue that both the BSP server implementation and the Bloop Config exporter needed to handle this on their own.
Now, this is handled in the IMHO right place, is properly shared and still overrideable by the user if the situation calls for it.

I recognized, that we don't support scalac plugin dependencies from in-project modules at the moment, but this can be added later if needed. Previously, it was enough to add the `module.jar()` to the `scalacPluginClassapth`, now it is also required to add the ```s"-Xplugin:{module.jar().path}"``` manually. Not to hard, but a future `scalacPluginModuleDeps` target could make this easier. As this also hits Mills build when we bootstrap our own project, that's why I did that change to `build.sc`.

Fix https://github.com/com-lihaoyi/mill/issues/1690
Fix https://github.com/com-lihaoyi/mill/issues/1691

Commits:

* Explicit handling of scalac plugins and related enablement options

We no longer generate an enablement option for each transitively resolved
plugin jar, but instead, only generated an -Xplugin option for each
explicitly given plugin. This will fix previously seen warnings for the
scalac compiler.

* More thorough review of options processing

* Removed special plugin handling from BloopImpl

* Fix build by explicitly adding plugin enablement

Until now, we inferred this option in the ZincWorker, but now, we infer it in ScalaModule from the ivy deps.
But here, we use a plugin from modules, which isn't supported out of the box.

* Fixed plugin path

* ScalaDoc updates

* Fixed plural form

* Fixed typos

* Imports cleanup

Pull request: https://github.com/com-lihaoyi/mill/pull/1705