Link Scala.js using Directory path when available (#1714)
This doesn't change any API but makes the Scala.js emit top level
exports.

## Motivation

Thinking about APIs is hard, so I tried to make the process of support the new `faskLinkJS` a two step approach. For now I migrated only the implementation to the new api using `DirectoryOutputPath` (in Scala. js 1.3+ only) but maintaed the `out.js` output path. It doesn't fail anymore if you try to export a module from the Scala.js code. At the same time it is not optimal since the auxiliary files are not returned by the `fastOpt` task, so are not considered for the tasks cache purposes.

Pull request: https://github.com/com-lihaoyi/mill/pull/1714