Added some MiMa problem filters
I have the feeling that our `mill.api.internal` annotation is ignored.
I added some ignore rules manually to get CI green, but I'm not sure if
it is actually safe to change the signature of these`bspXxx` methods.
It should, as they are only used internally, but my detailed knowledge of
trait encoding into classes/objects is a bit rusty.