Commands: Support tasks as parameters and make overriding usable (#1825)
Before this change, it was not possible to depend-on/call commands with parameters, when the parameter value came from another task. The typical workaround is to replace the parameter `T` by a `Task[T]`, but this still required to write an additional method and delegate to it from the command. 

This change allows now to directly use `Task[T]` for any supported `T`, so that no additional boilerplate code is necessary.

Example: 

```scala
import mill._
import mill.define._

object mod extends Module {

  def extra = T.input {
    "a"
  }

  def command1(arg: Task[String] = T.task()): Command[Unit] = T.command { 
    println(arg())
  }

  def command2(arg: String): Command[Unit] = T.command {
    command1(extra)()
  }

}
```

Fix https://github.com/com-lihaoyi/mill/issues/766

As a byproduct, this fixes als another issue we currently face when users try to override commands. It never really worked out of the box, as the parameters where not statically known. With the new ability to use tasks, we can finally override command and are still be able to call `super.command(task)()` to reuse the overridden command result.

Here is an example:

```scala
import mill._
import mill.define._

trait Mod extends Module {
  def command1(arg: Task[String] = T.task("hello")): Command[String] = T.command {
    val text = arg()
    println(text)
    text
  }
}

object mod extends Mod {
  override def command1(arg: Task[String] = T.task("")): Command[String] = T.command {
    super.command1(T.task("overridden " + arg()))()
  }
}
```

```sh
$ mill showNamed mod.command1 p1
[1/1] showNamed 
[1/1] showNamed > [1/2] mod.command1.overridden.ammonite.$file.build.Mod.command1 
overridden p1
[1/1] showNamed > [2/2] mod.command1 
{
  "mod.command1": "overridden p1"
}
```

I think this is so useful that we should make it a best practice for commands. We even should review existing commands and refactor them in the next major release cycle (0.11).

Pull request: https://github.com/com-lihaoyi/mill/pull/1825