Handle FileSystemException when creating an input stream (#1878)
When creating a `PathRef` of a path that contains socket files, which is for example the case if the path contains the `out` directory of Mill, Mill is failing because it can't read the socket file to calculate the content digest. We already skip digesting of unreadable files but somehow those socket files are not marked as unreadable. There is no performant and platform neutral way to detect socket files, so this PR just catches the typical `FileSystemException` at `InputStream` creation time and gracefully handles it (like other unreadable files) by just ignoring its content.

* Fix https://github.com/com-lihaoyi/mill/issues/1875
* Fix https://github.com/com-lihaoyi/mill/issues/823

Pull request: https://github.com/com-lihaoyi/mill/pull/1878