Bump zinc and account for diagnostic code
At the moment this is just testing a bit. I see you already have a test
PR to bump zinc in https://github.com/com-lihaoyi/mill/pull/1845 but one
of the things that this brings in is the changes to `Problem` I made in
https://github.com/sbt/sbt/pull/6874 that expose the diagnostic code of
the diagnostic coming from dotc. I have been doing some work on that on
the compiler side in https://github.com/lampepfl/dotty/pull/15565 and
wanted to try it out with Mill.

I tried to mimic the way you currently have it set up, so let me know if
it's not the direction you'd want to go. However, the idea here would be
that the diagnostic code is forwarded when diagnostics are published via
BSP so that Metals could then capture that code and know what code
actions to offer. You can see more of the big picture in https://github.com/lampepfl/dotty/issues/14904.