feat: add in a pluginArtifact helper
I've noticed myself (and copied from others) doing the same thing over
and over using the `scalaNativeBinaryVersion` help and then putting
together the artifact string ourselves, but this is easy to screw up if
you're not familiar with expected format. This just creates a helper
method that you can pass in the artifact name and Mill version and it
will put together the correct artifact name for you.