Fix issue with lazy val on Scala Native + Scala 3.2
Scala Native depends on the scala3-library `3.1.3`.
To work on Scala 3.2 you need to explicitly depend on scala3-library
for `3.2.0`.
In Scala.js it is a single artifact generated in the dotty repository,
so it's platformed. In Scala Native, however, we need to depend on the
jvm artifact, which is used by the compiler, so we patch
`scalaLibraryIvyDeps` to be `platformed = false`