Fixed inconsistency in args parser
Since we support separated arguments, we first separate all arguemnts and
later valdate and extract each separated group individualy.
When we get no args at all, we create no separated groups, hence we dont
have anything to validate and we miss to create a parse error for it.

This fix instead creates an empty first separated group, so we properly
validate it and detect the case of missing args.