Implemented workaround for resolving scoverage 1.x reporter API
The issue is, that Scoverage 1.x is no longer maintained,
but still the only option for projects on older Scala versions.
So, we try to keep it. Scala 2.13.9 and 2.13.10 are not directly
supported (as there are no full-version releases targeting the compiler).
But the reporting API is not compiler specific but also not shipped
independently, so we workaround that by resolving against 2.13.8.