Update Scala to 2.13.10 (#2055)
This one was a bit tricky, as just bumping the Scala version broke the
contrib.scoverage tests.

So I needed to implement a workaround for resolving the scoverage 1.x
reporter API.

The issue is, that Scoverage 1.x is no longer maintained, but still the
only option for projects on older Scala versions. So, we try to keep the
support for it. Scala 2.13.9 and 2.13.10 are not directly supported (as
there are no full-version releases targeting the compiler). But the
reporting API, which is not compiler specific, is not shipped
independently but inside the plugin jar. So we work around that by
resolving the plugin jar against Scala version 2.13.8 (the latest
supported version).

This also bumps Ammonite to 2.5.5 and Scoverage to 2.0.7.

Pull request: https://github.com/com-lihaoyi/mill/pull/2055