Switched to external mill-moduledefs dependency (#2068)
This is the second step of externalizing the `moduledefs` module.

The first step was to migrate the current `moduledefs` code into an
external project
[`mill-moduledefs`](https://github.com/com-lihaoyi/mill-moduledefs).

This second step now removes the `moduledefs` local module and refactors
the build setup to use the external one. In the external project, we
split the shared code and the scalac plugin code into two modules, to
respect the different use cases and technical requirements.

* The new `mill-moduledefs` aritfact contains the traits and classes,
which need to be on the classpath
* The `scalac-mill-moduledefs-plugin` artifact contains the scalac
plugin, which is distributed for the full compiler version and is only
used at compile-time

The overall goal is, to develop and maintain both projects (Mill and
mill-moduledefs) independently. It especially enables proper
distribution of the compiler plugin for each full version of the Scala
compiler. In addition, we can post-release the now independent
mill-moduledefs project for newer Scala versions without bumping the
version (if needed), so externally developed Mill plugin projects can
safely use newer Scala versions without risking compiler issues.

* Fixes https://github.com/com-lihaoyi/mill/issues/2035

Pull request: https://github.com/com-lihaoyi/mill/pull/2068