Catch even more coursier concurrency glitches, some consolidations (#2112)
This is an extension to our workarounds to coursier concurrency issues.

We handle some issues, already detected by courier and reported as a
result, as well as one which is not detected by coursier, so we
try-catch it and try to match based on the exception type and the
message.

As it is hard to reproduce such kind or issues, I created a GitHub
Actions workflow, which consistently failed without this very PR in step
"1. Run Mill with settings that should stress coursier". After I applied
this PR, all steps passed. That's not a 100-percent guarantee, but it
means, we're probably right on track.

```yaml
name: Stress test coursier downloads to reproduce concurrency issues

on:
  push:
  pull_request:

jobs:

  test-coursier:
    runs-on: ubuntu-latest

    steps:
      - uses: actions/checkout@v3
        with:
          fetch-depth: 0

      - uses: actions/setup-java@v3
        with:
          java-version: 11
          distribution: temurin

      - name: "Install local Mill into target and ivy cache"
        run: ./mill -i installLocal

      - name: "1. Cleanup courier cache"
        run: rm -r $HOME/.cache/coursier

      - name: "1. Run Mill with settings that should stress coursier"
        run: target/mill-release -i -j 20 -d __.prepareOffline

      - name: "2. Cleanup courier cache"
        run: rm -r $HOME/.cache/coursier

      - name: "2. Run Mill with settings that should stress coursier"
        run: target/mill-release -i -j 20 -d __.prepareOffline

```

Pull request: https://github.com/com-lihaoyi/mill/pull/2112