Update bloop-config to 1.5.4 and update used API
Bloop Config introduced some binary incompatible API change between 1.5.3 and 1.5.4.
I update the upickle JSON reader/writer setup and adapted to the new API.

I just forward a `None` to the source generators though, although, technically,
we probably should add a more sophisticated setup, as Mill is able to generate source code
and is also very aware of the fact it's generating it and if needs updates.
I simply left a TODO note in the code.