Support exoego fork of JsDom JsEnv (#2147)
My first approach to solve #2144 was to create a `Custom` `JsEnvConfig`
which would take a `className` and then the implementation would
instantiate that class with java reflection. It worked fine but didn't
support parameters or builders used by the JsEnv to create itself.
Since there aren't many custom `JsEnv`s out there (the only one I know
is [exoego](https://github.com/exoego/scala-js-env-jsdom-nodejs)'s fork
of jsdom-nodejs, I'm proposing of supporting it officially in Mill.
This PR also changes the worker classpath to only download and load the
jsEnv dependencies that are needed. This skips downloading artifacts and
loading classes for jsEnvs that aren't used, like this new one for
people not using it, or the deprecated phantom jsEnv.
 
Testing this properly means installing `jsdom` which can only be done in
the root because of https://github.com/com-lihaoyi/mill/issues/1036 (to
my understanding). I tested it manually in scratch but I needed to have
a global `package.json`

Pull request: https://github.com/com-lihaoyi/mill/pull/2147