Always give included files a unique foreign module path (#2132)
With this change, we always give included files a unique foreign module
path. This makes the `dest` path more consistent and robust.

Unfortunately, this slightly changes the encoding of foreign modules
names. This is no issue when writing `.sc` files, but it will be visible
in the progress output and in BSP client and IntelliJ Idea. E.g. the
name of an included file in directory `proj1` containing a module
`proj1`, will now have the module name
"foreign-modules.proj1.build.proj1", whereas previously it was
"foreign-modules.proj1.proj1".

* Fix https://github.com/com-lihaoyi/mill/issues/2130

Pull request: https://github.com/com-lihaoyi/mill/pull/2132