Rudimentary cache for zinc compilation problems
In it's incremental nature, Zinc does only report problems for re-compiled files,
which leads to situations, where you only see a glimpse of the overall compiler warnings.

With this change, I keep a record of all reported problems, so they can be later re-shown to the user.
The cache is dropping all recorded problems for files, which were recompiled.