Fix concurrency issue with test event reporting (Backport of #2314)
An intermittent failure occurs when running tests with `-j 0 ...` on a
machine with plenty of cores. This is due to concurrent use of
scala.collection.mutable.Buffer.

This PR uses a concurrent collection to allow concurrent tests to append
events safely.

Original Pull Request: https://github.com/com-lihaoyi/mill/pull/2314