Reworked validated PathRef design
We now check the validity of `PathRef`s upon upickle reading
and throw an exception if it is invalid. Beside a much simpler
implementation, this has also the benefit, that we can detect
deeply nested invalid `PathRef`s.

Failing deserialization is already handled in Evaluator and
the cached value is automatically dropped.

A `PathRef` can have three revalidation states, never, once and always.
Once means, that the revalidation state is changed to never when it
is successfully read from json.