Support automatic `PathRef` validation for cached targets (#2328)
We now check the validity of `PathRef`s upon upickle reading
and throw an exception if it is invalid. This also works, if the
`PathRef` is deeply nested in the result. A `PathRef` is valid,
when it produced the same `sig` on the current filesystem.

Failing deserialization is already handled in Evaluator and
the cached value is automatically dropped. In such a case we also
write a debug level log entry.

A `PathRef` can have three revalidation states, never, once and always.
Once means, that the revalidation state is changed to never when it
is successfully read from json.

Why is this needed? Normally, `PathRef`s referring to files under
`T.dest` do not need extra validation, as the target specific directory
is considered as Mill managed storage. But if a `PathRef`s refers to a
file or directory outside the Mill cache directory, it can become out of
sync. The most prominent example is `JavaModule.resolvedIvyDeps`, which
resolves and downloads the `ivyDeps` into an external storage location
(`~/.cache/coursier`). If a user clears that cache storage, the cached
result is stale, yet Mill does not recognize it. By marking these
`PathRef`s to re-validate, the evaluator will check that the external
cache storage is as expected and the returned `PathRef`s are valid.
Otherwise, it will be re-resolved.

Pull request: https://github.com/com-lihaoyi/mill/pull/2328